Added option to build for specific commit

+++ build.sh
@@ -12,6 +12,20 @@ NC='\033[0m' # No Color
 # https://golang.org/doc/install/source#environment
 GOARCH_ALL="386 amd64 arm arm64 ppc64 ppc64le mips mipsle mips64 mips64le riscv64 s390x"
 
+COMMIT="HEAD"
+ARM=""
+
+while getopts :a:c: option; do
+    case "${option}" in
+        a)  case "$OPTARG" in 5|6|7) ARM="$ARM $OPTARG";; *) echo "Unsupported ARM version $OPTARG"; exit 2;; esac;;
+        c)  COMMIT="$OPTARG";;
+        *)  echo "Unknown option -${OPTION}"; exit 2;;
+    esac
+done
+shift $((OPTIND-1))
+
+[ -z "$ARM" ] && ARM="7 6 5"
+
 if [ $# -eq 0 ]; then
     set -- "arm"
 elif [ "$1" = "all" ]; then
@@ -28,8 +42,11 @@ fi
 
 gitreset() {
     git fetch
-    git reset --hard HEAD
-    git merge '@{u}'
+    git reset --hard $COMMIT || exit 2
+    if [ "$COMMIT" = "HEAD" ]; then
+        git merge '@{u}'
+        git reset --hard HEAD
+    fi
 }
 
 checkSHA256() {
@@ -66,6 +83,25 @@ wireguard() {
     cd ${REPO_ROOT_DIR}/wireguard-go
     echo "${GREEN} -> Pulling latest wireguard-go repository...${GREY}"
     gitreset
+
+    case $(git rev-parse --short HEAD) in
+        3b95c81) # 2022-02-13
+            UNMODIFIED_main="8c58063f67f63d91d64dec6072cf728a3449a1b263c99074c850db0a630f6058"
+            UNMODIFIED_makefile="f59c6fbbe54c2d194207ef93bdb27ab69a4f67efd26f147f3a0c60268ebaf57c"
+            UNMODIFIED_queueconstants_default="470364f455a6637f061cf6363929e8977f7872b43fd6f5ea008e223671b5330c"
+            MODIFIED_main="18c2174e0a22c3e9ac6fd5d077ed8e7e97ec401f3207bf45e5b590964fc4ace4"
+            MODIFIED_makefile="0b650215e15b92e0b185fc56bc517390cb35e1d36af0dea09920b84c568065c1"
+            MODIFIED_queueconstants_default="7a7fdce9ae60633d82c54749edffffa4de9a30f0352009a11383c30d8c2654b3"
+            ;;
+        bb745b2) # 2021-09-27
+            UNMODIFIED_main="1889250813d3fc9e4538e669b4fe86fd2caa4949094be06033e6a5c0eb6deb29"
+            UNMODIFIED_makefile="f59c6fbbe54c2d194207ef93bdb27ab69a4f67efd26f147f3a0c60268ebaf57c"
+            UNMODIFIED_queueconstants_default="461802f0fac24977a6164ac96b47b59740c506ed124c39a9e434493889384f28"
+            MODIFIED_main="2cf3bcb37be0f4e4e58ccc416ba16a6bec61261f12271afa7c9aedceacf51589"
+            MODIFIED_makefile="c04d9998ae41f016319fb49cc7ffe4955c016368b880814980335fce48b961a2"
+            MODIFIED_queueconstants_default="228c3ed2e4851c988d6a0e4837d18d26f32880331beaa723f13d9aa27dd2be51"
+            ;;
+    esac
 
     if checkSHA256 main.go $UNMODIFIED_main && checkSHA256 device/queueconstants_default.go $UNMODIFIED_queueconstants_default; then
         echo "${GREEN} -> Removing 'first class kernel support' intercept from main.go...${GREY}"
@@ -190,7 +226,7 @@ if wg && wireguard; then
     for a in $@; do
         export GOARCH="$a"
         if [ "$a" = "arm" ]; then
-            for v in 7 6 5; do
+            for v in $ARM; do
                 export GOARM="$v"
                 build_release "${a}v${v}"
                 unset GOARM
